Resolved compile errors after merge. Moved radius from Dimension to Projectile.
